+++ src/date-picker2/picker.jsx
@@ -19,14 +19,25 @@ const { Popup } = Overlay;
 const { renderNode } = func;
 const { pickProps, pickOthers } = obj;
 
+/**
+ * 判断值是否改变
+ * @param {dayjs.ConfigType}} newValue
+ * @param {dayjs.ConfigType} oldValue
+ * @returns {boolean}
+ */
 function isValueChanged(newValue, oldValue) {
     return Array.isArray(newValue)
         ? isValueChanged(newValue[0], oldValue && oldValue[0]) || isValueChanged(newValue[1], oldValue && oldValue[1])
         : newValue !== oldValue && !datejs(newValue).isSame(oldValue);
 }
 
-// 返回日期字符串
-function getInputValue(value, fmt) {
+/**
+ * 获取输入框值
+ * @param {*} value 日期值
+ * @param {string | funtion} fmt 日期格式
+ * @returns {string | string[]}
+ */
+function fmtValue(value, fmt) {
     const formater = (v, idx) => {
         if (Array.isArray(fmt)) {
             fmt = fmt[idx];
@@ -37,27 +48,50 @@ function getInputValue(value, fmt) {
     return Array.isArray(value) ? value.map((v, idx) => formater(v, idx)) : formater(value);
 }
 
-// 无效值转为 null
-// undefined 值表示当前时间
+/**
+ * 日期检验：无效值返回 null
+ * @param {dayjs.ConfigType} value
+ * @returns {Dayjs | null}
+ */
 function checkDate(value) {
-    // 因为datejs(undefined) === datejs()
-    // 但是这里期望的是一个空值
+    /**
+     * 因为 datejs(undefined) 表示当前时间
+     * 但是这里期望的是一个空值，即用户不输入值的时候显示为空
+     */
     if (value === undefined) {
         value = null;
     }
+
     value = datejs(value);
     return value.isValid() ? value : null;
 }
 
+/**
+ * Range 日期检验
+ * @param {dayjs.ConfigType[]} value 日期值
+ * @param {number} inputType 输入框类型：开始时间输入框/结束时间输入框
+ * @param {boolean} disabled 是否禁用
+ * @param {boolean} strictly 是否严格校验：严格模式下不允许开始时间大于结束时间，在显示确认按键的，用户输入过程可不严格校验
+ * @returns {Dayjs[] | null[]}
+ */
 function checkRangeDate(value, inputType, disabled, strictly = true) {
     const [begin, end] = Array.isArray(value) ? [0, 1].map(i => checkDate(value[i])) : [null, null];
-    const _disabled = Array.isArray(disabled) ? disabled : [disabled, disabled];
+    const [disabledBegin, disabledEnd] = Array.isArray(disabled) ? disabled : [disabled, disabled];
 
-    // 严格模式下
-    // 如果开始时间在结束时间之后 清空另一个时间
-    // 如果另一个时间处于禁用状态 清除当前时间
+    /**
+     * 需要清除其中一个时间，优先清除结束时间，下面情况清除开始时间：
+     * 1. 结束时间被 disabled 而开始时间没有被 disabled
+     * 2. 开始时间和结束时间都没被 disabled 且当前正在输入是结束时间
+     */
     if (strictly && begin && end && begin.isAfter(end)) {
-        return inputType === DATE_INPUT_TYPE.BEGIN && !_disabled[1] ? [begin, null] : [null, end];
+        if (
+            (!disabledBegin && disabledEnd) ||
+            (!disabledBegin && !disabledBegin && inputType === DATE_INPUT_TYPE.END)
+        ) {
+            return [null, end];
+        }
+
+        return [begin, null];
     }
 
     return [begin, end];
@@ -109,9 +143,20 @@ class Picker extends React.Component {
         placeholder: SharedPT.placeholder,
         disabled: SharedPT.disabled,
         inputReadOnly: SharedPT.readOnly,
+        /**
+         * 日期显示格式
+         */
         format: SharedPT.format,
         label: PT.node,
         separator: PT.node,
+        /**
+         * 输出格式：控制 onChange、onOk 事件的输出值格式
+         *  - string 类型：根据时间格式进行转换
+         *  - function 类型：((date: Dayjs, dateStr: string) => any)
+         *
+         * @version 1.23
+         */
+        outputFormat: SharedPT.format,
 
         // popup
         followTrigger: PT.bool,
@@ -165,7 +210,7 @@ class Picker extends React.Component {
 
         this.state = {
             value,
-            inputValue: getInputValue(value, format),
+            inputValue: fmtValue(value, format),
             curValue: value, // 当前状态值
             ...this.state,
         };
@@ -175,13 +220,32 @@ class Picker extends React.Component {
 
     static getDerivedStateFromProps(props, state) {
         const { type, showTime, showOk, disabled, format } = props;
-
         const isRange = type === DATE_PICKER_TYPE.RANGE;
 
         let newState = {
             isRange,
             showOk: !!(showOk || showTime),
         };
+
+        /**
+         * 如果当前输入框可能被 disabled
+         * 且另一个输入框非 disabled 则切换到另一个输入框
+         */
+        if (isRange) {
+            let { inputType } = state;
+
+            const _disabled = Array.isArray(disabled) ? disabled : [disabled, disabled];
+
+            if (_disabled[inputType]) {
+                const otherType = switchInputType(state.inputType);
+
+                if (!_disabled[otherType]) {
+                    inputType = otherType;
+                }
+            }
+
+            newState.inputType = inputType;
+        }
 
         if ('value' in props) {
             const value = isRange ? checkRangeDate(props.value, state.inputType, disabled) : checkDate(props.value);
@@ -191,7 +255,7 @@ class Picker extends React.Component {
                     ...newState,
                     value,
                     curValue: value,
-                    inputValue: getInputValue(value, format),
+                    inputValue: fmtValue(value, format),
                 };
             }
         }
@@ -203,6 +267,9 @@ class Picker extends React.Component {
         [this.clearTimeoutId, this.timeoutId].forEach(id => id && clearTimeout(id));
     }
 
+    /**
+     * 获取初始值
+     */
     getInitValue = () => {
         const { props } = this;
         const { type, value, defaultValue } = props;
@@ -214,6 +281,12 @@ class Picker extends React.Component {
         return this.checkValue(val);
     };
 
+    /**
+     * 获取 RangePicker 输入框初始输入状态
+     * @returns {Object} inputState
+     * @returns {boolean} inputState.justBeginInput 是否初始输入
+     * @returns {number} inputState.inputType 当前输入框
+     */
     getInitRangeInputState = () => {
         return {
             justBeginInput: this.isEnabled(),
@@ -221,11 +294,17 @@ class Picker extends React.Component {
         };
     };
 
-    // 判断弹层是否显示
+    /**
+     * 处理点击文档区域导致的弹层收起逻辑
+     * @param {boolean} visible 是否可见
+     * @param {string} type 事件类型
+     */
     handleVisibleChange = (visible, type) => {
         if (type === 'docClick') {
-            visible || this.handleChange(this.state.curValue, 'VISIBLE_CHANGE');
-
+            // 弹层收起 触发 Change 逻辑
+            if (!visible) {
+                this.handleChange(this.state.curValue, 'VISIBLE_CHANGE');
+            }
             this.onVisibleChange(visible);
         }
     };
@@ -293,8 +372,10 @@ class Picker extends React.Component {
     };
 
     handleClear = () => {
-        // 清空输入之后 input组件内部会让第二个输入框获得焦点
-        // 所以这里需要设置setTimeout才能让第一个input获得焦点
+        /**
+         * 清空输入之后 input 组件内部会让第二个输入框获得焦点
+         * 所以这里需要设置 setTimeout 才能让第一个 input 获得焦点
+         */
         this.clearTimeoutId = setTimeout(() => {
             this.handleInputFocus(0);
         });
@@ -328,13 +409,28 @@ class Picker extends React.Component {
             : !disabled;
     };
 
+    onKeyDown = e => {
+        switch (e.keyCode) {
+            case KEYCODE.ENTER: {
+                const { inputValue } = this.state;
+                this.onClick();
+                this.handleChange(inputValue, 'KEYDOWN_ENTER');
+                break;
+            }
+            default:
+                return;
+        }
+    };
+
     handleChange = (v, eventType) => {
         const { format } = this.props;
         const { isRange, showOk, value } = this.state;
         const forceEvents = ['KEYDOWN_ENTER', 'CLICK_OK', 'CLICK_PRESET', 'INPUT_CLEAR'];
 
-        // 在显示确认按键且关闭弹层的时候 将当前值设置回确认值
-        // 在有确认按键的时候 需要点击确认值才生效
+        /**
+         * 在显示确认按键且关闭弹层的时候 将当前值设置回确认值
+         * 在有确认按键的时候 需要点击确认值才生效
+         */
         if (showOk && eventType === 'VISIBLE_CHANGE') {
             v = value;
         } else {
@@ -344,7 +440,7 @@ class Picker extends React.Component {
 
         this.setState({
             curValue: v,
-            inputValue: getInputValue(v, format),
+            inputValue: fmtValue(v, format),
         });
 
         if (!showOk || forceEvents.includes(eventType)) {
@@ -359,19 +455,6 @@ class Picker extends React.Component {
         }
     };
 
-    onKeyDown = e => {
-        switch (e.keyCode) {
-            case KEYCODE.ENTER: {
-                const { inputValue } = this.state;
-                this.onClick();
-                this.handleChange(inputValue, 'KEYDOWN_ENTER');
-                break;
-            }
-            default:
-                return;
-        }
-    };
-
     onChange = v => {
         const { value } = this.state;
         const { format } = this.props;
@@ -381,7 +464,7 @@ class Picker extends React.Component {
             if ('value' in this.props) {
                 this.setState({
                     curValue: value,
-                    inputValue: getInputValue(value, format),
+                    inputValue: fmtValue(value, format),
                 });
             } else {
                 v = this.checkValue(v);
@@ -389,11 +472,11 @@ class Picker extends React.Component {
                 this.setState({
                     value: v,
                     curValue: v,
-                    inputValue: getInputValue(v, format),
+                    inputValue: fmtValue(v, format),
                 });
             }
 
-            func.invoke(this.props, 'onChange', [v, getInputValue(v, format)]);
+            func.invoke(this.props, 'onChange', this.getOutputArgs(v));
         }
 
         this.onVisibleChange(false);
@@ -401,8 +484,9 @@ class Picker extends React.Component {
 
     onOk = () => {
         const { inputValue } = this.state;
+        const checkedValue = this.checkValue(inputValue);
 
-        const result = func.invoke(this.props, 'onOk', [this.checkValue(inputValue), inputValue]);
+        const result = func.invoke(this.props, 'onOk', this.getOutputArgs(checkedValue));
 
         result !== false && this.handleChange(inputValue, 'CLICK_OK');
     };
@@ -425,6 +509,21 @@ class Picker extends React.Component {
             this.onVisibleChange(true);
             this.handleInputFocus(inputType);
         }
+    };
+
+    /**
+     * 获取 `onChange` 和 `onOk` 方法的输出参数
+     * @param {Dayjs} value
+     * @returns 默认返回 `Dayjs` 实例和 `format` 格式化的值
+     *          如果传了了 `outputFormat` 属性则返回 `outputFormat` 格式化的值
+     */
+    getOutputArgs = value => {
+        const { outputFormat, format } = this.props;
+
+        if (outputFormat) {
+            return [fmtValue(value, outputFormat)];
+        }
+        return [value, fmtValue(value, format)];
     };
 
     getCurrentAlign = ({ align }) => {