+++ src/table/base.jsx
@@ -411,6 +411,10 @@ class Table extends React.Component {
                 if (child) {
                     const props = { ...child.props };
 
+                    if (child.ref) {
+                        props.ref = child.ref;
+                    }
+
                     if (
                         !(
                             child &&