+++ src/table/base/header.jsx
@@ -39,6 +39,12 @@ export default class Header extends React.Component {
 
     getCellRef = (i, j, cell) => {
         this.props.headerCellRef(i, j, cell);
+
+        const { columns } = this.props;
+        const columnProps = columns[i] && columns[i][j];
+        if (columnProps && columnProps.ref && typeof columnProps.ref === 'function') {
+            columnProps.ref(cell);
+        }
     };
 
     onSort = (dataIndex, order, sort) => {
@@ -99,12 +105,14 @@ export default class Header extends React.Component {
                     __normalized,
                     lock,
                     cellStyle,
+                    wordBreak,
                     ...others
                 } = col;
 
                 className = classnames({
                     [`${prefix}table-header-node`]: true,
                     [`${prefix}table-header-resizable`]: resizable,
+                    [`${prefix}table-word-break-${wordBreak}`]: !!wordBreak,
                     [className]: className,
                 });
                 let attrs = {},